requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
python-multipart==0.0.9
psycopg2-binary==2.9.9
python-dotenv==1.0.1
pypdf==4.2.0
python-docx==1.1.2
requests==2.32.3
